pyproject.toml
[tool.poetry]
name = "bookstore1"
version = "0.1.0"
description = "Bookstore API"
authors = ["kevyn 60 <kevynsousa606@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
django-rest-framework = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
